requirements.txt
flask
flask-cors
python-dotenv
openai
transformers
sentencepiece
torch
